refactored search and section selector
