+++ src/components/schedule/Line.js
@@ -1,0 +1,3 @@
+export default Line = props => (
+  <hr width="99.7%" className="schedline" style={{top: props.y + "%"}}/>
+)
