+++ src/components/schedule_components.js
@@ -1,0 +1,4 @@
+import React, { Component } from 'react';
+
+
+// ReactDOM.render(<SchedGrid schedData ={[]}/>, document.querySelector("#Schedule"));
